+++ czkawka_core/src/lib.rs
@@ -6,7 +6,6 @@ pub mod tools;
 pub mod re_exported {
     pub use fast_image_resize::FilterType as FirFilterType;
     pub use image_hasher::{FilterType, HashAlg};
-    pub use vid_dup_finder_lib::Cropdetect;
 }
 
 pub const CZKAWKA_VERSION: &str = env!("CARGO_PKG_VERSION");