+++ krokiet/src/connect_directories_changes.rs
@@ -1,3 +1,5 @@
+use std::collections::HashSet;
+
 use rfd::FileDialog;
 use slint::{ComponentHandle, Model, ModelRc, VecModel};
 
@@ -117,10 +119,10 @@ fn connect_add_files(app: &MainWindow) {
 
 fn add_included_paths(settings: &Settings, folders: &[String]) {
     let old_folders = settings.get_included_paths_model();
-    let old_folders_path = old_folders.iter().map(|x| x.path.to_string()).collect::<Vec<_>>();
+    let old_folders_path: HashSet<String> = old_folders.iter().map(|x| x.path.to_string()).collect();
     let mut new_folders = old_folders.iter().collect::<Vec<_>>();
 
-    let filtered_folders = folders.iter().filter(|x| !old_folders_path.contains(x)).collect::<Vec<_>>();
+    let filtered_folders = folders.iter().filter(|x| !old_folders_path.contains(x.as_str())).collect::<Vec<_>>();
 
     for x in &mut new_folders {
         x.focused_row = false;
@@ -140,10 +142,10 @@ fn add_included_paths(settings: &Settings, folders: &[String]) {
 
 pub(crate) fn add_excluded_paths(settings: &Settings, folders: &[String]) {
     let old_folders = settings.get_excluded_paths_model();
-    let old_folders_path = old_folders.iter().map(|x| x.path.to_string()).collect::<Vec<_>>();
+    let old_folders_path: HashSet<String> = old_folders.iter().map(|x| x.path.to_string()).collect();
     let mut new_folders = old_folders.iter().collect::<Vec<_>>();
 
-    let filtered_folders = folders.iter().filter(|x| !old_folders_path.contains(x)).collect::<Vec<_>>();
+    let filtered_folders = folders.iter().filter(|x| !old_folders_path.contains(x.as_str())).collect::<Vec<_>>();
 
     for x in &mut new_folders {
         x.focused_row = false;