+++ Changelog.md
@@ -28,6 +28,7 @@ The project will remain available in the repository. For some time I will ensure
 - Fixed a bug where relative symlinks were resolved without considering the parent directory - [#1900](https://github.com/qarmin/czkawka/pull/1900)
 - Fixed a prehash cache bypass that caused full hash computation on some files during a second scan - [#1907](https://github.com/qarmin/czkawka/pull/1907)
 - Extended integration tests, to prevent regressions in the future - [#1919](https://github.com/qarmin/czkawka/pull/1919)
+- Fix unstable prehash if `read` function returned less bytes than requested - [#]()
 
 ### CLI